requirements.txt
# Data Pipeline Dependencies
# Essential packages only

pandas>=1.4.0
pandera>=0.26.0
pyyaml>=6.0
pyarrow>=10.0.0